ref #12 - failing test case
load order of CGI and CGI::Fast can cause import pragmas in CGI to
be overwritten, which is probably not good behaviour. bisect shows
this was introduced in 53651f2. the fix maybe just do document this
behaviour given its age

squash this commit with the fix (or remove the test if the fix is
just documentation)