fix #19 - ensure hooks are passed to CGI.pm constructor
A CGI upload hook is used, among other reasons, to prevent sensitive
user data from being written to disk.

The hook must be provided to the CGI constructor, but CGI::Fast
mistakenly identifies the hook as an initializer, and this prevents a
hook from being used.

this came out as a question on stackoverflow[1], which is repeated here
should it be lost to the web-mists of time:

---------------------------------

CGI::Fast without tempfiles for upload

I wish to upload some relatively sensitive data to a server which should
not be stored on disk in plaintext - I see it as far too easy to
intercept a temporary file and copy the contents, plus there would be
the issues of securely deleting the file against restores/journal
entries, etc. I realise I could theoretically do this client-side, but
prefer a server-side solution so an unusual browser does not mangle the
upload.

Perl CGI::upload_hook() seemed like a perfect solution, with encryption
being applied on arrival. Plain CGI is deprecated, but using CGI::Fast,
the hook code is never reached, whether I put the call inside or outside
the loop.

    use Sys::Syslog;
    use CGI::Fast;
    sub hook {
	my ( $filename, $buffer, $bytes_read, $data ) = @_;
	syslog('notice',"Hook used!");
    }

    #CGI::upload_hook(\&hook,'',0); # No good here either
    while (my $q = CGI::Fast->new()) {
	CGI::upload_hook(\&hook,'data',0); # No effect
	syslog('notice',"Got a request!");
	# (Access control code cut)
	print $q->header('text/plain','200 OK');
	print (Dumper($q));
    }

hook() is never called. CGI::Fast->new() does not accept such hook code
in the way that CGI->new() does.

asked Mar 25 at 22:22 somloigaluska

=================================

Once the CGI object is created, it's too late to set the hook.

First of all, you need the following fix to CGI::Fast::new:

    sub new {
	my ($self, @args) = @_;

	if (!$args[0] || ( ref($args[0]) && UNIVERSAL::isa($args[0], 'CODE') && !$args[3] ) ) {
	    $Ext_Request ||= _create_fcgi_request( $in_fh,$out_fh,$err_fh );
	    return undef unless $Ext_Request->Accept >= 0;
	}

	CGI->_reset_globals;
	$self->_setup_symbols(@CGI::SAVED_SYMBOLS) if @CGI::SAVED_SYMBOLS;
	return $CGI::Q = $self->SUPER::new(@args);
    }

If you want to use a monkey patch instead of getting the maintainer to
fix the bug, you can use the following:

    BEGIN {
	package CGI::Fast;

	my ( $in_fh,$out_fh,$err_fh );

	my $file_handles = sub {
	    my ($self, $handles) = @_;

	    if ( ref( $handles ) eq 'HASH' ) {
		$in_fh  = delete( $handles->{fcgi_input_file_handle} );
		$out_fh = delete( $handles->{fcgi_output_file_handle} );
		$err_fh = delete( $handles->{fcgi_error_file_handle} );
	    }
	};

	my $new = sub {
	    my ($self, @args) = @_;

	    if (!$args[0] || ( ref($args[0]) && UNIVERSAL::isa($args[0], 'CODE') && !$args[3] ) ) {
		$Ext_Request ||= _create_fcgi_request( $in_fh,$out_fh,$err_fh );
		return undef unless $Ext_Request->Accept >= 0;
	    }

	    CGI->_reset_globals;
	    $self->_setup_symbols(@CGI::SAVED_SYMBOLS) if @CGI::SAVED_SYMBOLS;
	    return $CGI::Q = $self->SUPER::new(@args);
	};

	no warnings qw( redefine );
	*file_handles = $file_handles;
	*new = $new;
    }

If you want to fix it yourself instead of getting the maintainer to fix
the bug, and you don't use CGI::Fast->file_handles, you can make do with
the following monkey patch:

    BEGIN {
	package CGI::Fast;

	my $file_handles = sub {
	    die "Not supported.\n";
	};

	my $new = sub {
	    my ($self, @args) = @_;

	    if (!$args[0] || ( ref($args[0]) && UNIVERSAL::isa($args[0], 'CODE') && !$args[3] ) ) {
		$Ext_Request ||= _create_fcgi_request( undef, undef, undef );
		return undef unless $Ext_Request->Accept >= 0;
	    }

	    CGI->_reset_globals;
	    $self->_setup_symbols(@CGI::SAVED_SYMBOLS) if @CGI::SAVED_SYMBOLS;
	    return $CGI::Q = $self->SUPER::new(@args);
	};

	no warnings qw( redefine );
	*file_handles = $file_handles;
	*new = $new;
    }

After either monkey patch, you can use the following:

    while (1) {
	my $q = CGI::Fast->new(\&hook, 'data', 0)
	   or die("Couldn't Accept FCGI request\n");

	syslog('notice',"Got a request!");
	# (Access control code cut)
	print $q->header('text/plain','200 OK');
    }

answered Mar 26 at 4:12 ikegami

---------------------------------

bump VERSION and Changes for CPAN release

[1] https://stackoverflow.com/questions/55347257/cgifast-without-tempfiles-for-upload